+++ AmazonChimeSDK/AmazonChimeSDK/analytics/EventName.swift
@@ -14,8 +14,10 @@ import Foundation
     case audioInputFailed
     /// The camera selection failed.
     case videoInputFailed
-    /// The microphone or camera device access failed.
-    case deviceAccessFailed
+    /// The microphone access failed.
+    case audioAccessFailed
+    /// The camera access failed.
+    case videoAccessFailed
     /// The meeting will start.
     case meetingStartRequested
     /// The meeting started.
@@ -37,8 +39,10 @@ import Foundation
             return "audioInputFailed"
         case .videoInputFailed:
             return "videoInputFailed"
-        case .deviceAccessFailed:
-            return "deviceAccessFailed"
+        case .audioAccessFailed:
+            return "audioAccessFailed"
+        case .videoAccessFailed:
+            return "videoAccessFailed"
         case .meetingStartRequested:
             return "meetingStartRequested"
         case .meetingStartSucceeded:
@@ -62,8 +66,10 @@ import Foundation
             return .audioInputFailed
         case "videoInputFailed":
             return .videoInputFailed
-        case "deviceAccessFailed":
-            return .deviceAccessFailed
+        case "audioAccessFailed":
+            return .audioAccessFailed
+        case "videoAccessFailed":
+            return .videoAccessFailed
         case "meetingStartRequested":
             return .meetingStartRequested
         case "meetingStartSucceeded":