+++ AmazonChimeSDK/AmazonChimeSDK/ingestion/IngestionEventConverter.swift
@@ -37,9 +37,14 @@ private typealias EAName = EventAttributeName
             audioErrorStr = String(describing: audioError)
         }
         
-        var deviceAccessErrorStr: String?
-        if let deviceAccessError = eventAttributes[EventAttributeName.deviceAccessFailedError] as? Error {
-            deviceAccessErrorStr = String(describing: deviceAccessError)
+        var audioAccessErrorStr: String?
+        if let audioAccessError = eventAttributes[EventAttributeName.audioAccessError] as? Error {
+            audioAccessErrorStr = String(describing: audioAccessError)
+        }
+        
+        var videoAccessErrorStr: String?
+        if let videoAccessError = eventAttributes[EventAttributeName.videoAccessError] as? Error {
+            videoAccessErrorStr = String(describing: videoAccessError)
         }
 
         var attributes = [String:AnyCodable]()
@@ -58,7 +63,8 @@ private typealias EAName = EventAttributeName
         attributes[EAName.retryCount.description] = AnyCodable(eventAttributes[EAName.retryCount])
         attributes[EAName.videoInputError.description] = AnyCodable(videoErrorStr)
         attributes[EAName.audioInputError.description] = AnyCodable(audioErrorStr)
-        attributes[EAName.deviceAccessFailedError.description] = AnyCodable(deviceAccessErrorStr)
+        attributes[EAName.audioAccessError.description] = AnyCodable(audioAccessErrorStr)
+        attributes[EAName.videoAccessError.description] = AnyCodable(videoAccessErrorStr)
         
         clientConfig.metadataAttributes.forEach({ (key: String, value: Any) in
             attributes[key] = AnyCodable(value)
@@ -168,7 +174,8 @@ private typealias EAName = EventAttributeName
                                 retryCount: meetingEvent.getRetryCount(),
                                 videoInputErrorMessage: meetingEvent.getVideoInputErrorMessage(),
                                 audioInputErrorMessage: meetingEvent.getAudioInputErrorMessage(),
-                                deviceAccessErrorMessage: meetingEvent.getDeviceAccessErrorMessage(),
+                                audioAccessErrorMessage: meetingEvent.getAudioAccessErrorMessage(),
+                                videoAccessErrorMessage: meetingEvent.getVideoAccessErrorMessage(),
                                 ttl: dirtyMeetingEventTtl)
     }