+++ AmazonChimeSDK/AmazonChimeSDK/internal/audio/DefaultAudioClientObserver.swift
@@ -51,16 +51,6 @@ class DefaultAudioClientObserver: NSObject, AudioClientDelegate {
     public func audioClientStateChanged(_ audioClientState: audio_client_state_t, status: audio_client_status_t) {
         let newAudioStatus = Converters.AudioClientStatus.toMeetingSessionStatusCode(status: status)
         let newAudioState = Converters.AudioClientState.toSessionStateControllerAction(state: audioClientState, status: newAudioStatus)
-
-        if(newAudioStatus == .audioInputDeviceNotResponding) {
-            eventAnalyticsController.publishEvent(name: .deviceAccessFailed, attributes: [
-                EventAttributeName.deviceAccessFailedError: DeviceError.audioInputDeviceNotRespondingError
-            ])
-        } else if(newAudioStatus == .audioOutputDeviceNotResponding) {
-            eventAnalyticsController.publishEvent(name: .deviceAccessFailed, attributes: [
-                EventAttributeName.deviceAccessFailedError: DeviceError.audioOutputDeviceNotRespondingError
-            ])
-        }
         
         if newAudioStatus == .unknown {
             logger.info(msg: "AudioClient State rawValue: \(audioClientState.rawValue) Unknown Status rawValue: \(status.rawValue)")