+++ AmazonChimeSDK/AmazonChimeSDKTests/analytics/EventAttributeNameTests.swift
@@ -36,6 +36,7 @@ class EventAttributeNameTests: XCTestCase {
         XCTAssertEqual(EventAttributeName.retryCount.description, "retryCount")
         XCTAssertEqual(EventAttributeName.audioInputError.description, "audioInputError")
         XCTAssertEqual(EventAttributeName.videoInputError.description, "videoInputError")
-        XCTAssertEqual(EventAttributeName.deviceAccessFailedError.description, "deviceAccessFailedError")
+        XCTAssertEqual(EventAttributeName.audioAccessError.description, "audioAccessError")
+        XCTAssertEqual(EventAttributeName.videoAccessError.description, "videoAccessError")
     }
 }